<DE2 & Android>Command, CmdScheduler Modify the command function play to take id as parameter instead of song name

+++ Android_Workspace/ece381/src/com/example/ece381/CmdScheduler.java
@@ -29,7 +29,7 @@ public class CmdScheduler {
 			cmd = cmds.poll();
 			switch(cmd.getCmdIndex()) {
 			case 1:
-				Command.play(cmd.getParameterList().get(0), Integer.valueOf(cmd.getParameterList().get(1)).intValue());
+				Command.play(Integer.valueOf(cmd.getParameterList().get(0)).intValue(), Integer.valueOf(cmd.getParameterList().get(1)).intValue());
 				Log.i("COMMAND", "The Song "+cmd.getParameterList().get(0)+" is playing at "+cmd.getParameterList().get(1));
 				break;
 			case 2:
